Split main.rs file into multiple modules
