+++ compiler/cli/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod stdin;
+pub mod stdout;
